remove rollup-core dependency from contracts
