+++ packages/contracts/package.json
@@ -40,7 +40,6 @@
     "@eth-optimism/core-db": "^0.0.1-alpha.25",
     "@eth-optimism/core-utils": "^0.0.1-alpha.25",
     "@eth-optimism/solc-transpiler": "^0.0.1-alpha.27",
-    "@eth-optimism/rollup-core": "^0.0.1-alpha.28",
     "@nomiclabs/buidler": "^1.3.8",
     "@nomiclabs/buidler-ethers": "^2.0.0",
     "@nomiclabs/buidler-solpp": "^1.3.3",
