+++ packages/contracts/test/contracts/ovm/L2ExecutionManager.spec.ts
@@ -4,7 +4,6 @@ import '../../setup'
 import { ethers } from '@nomiclabs/buidler'
 import { add0x, getLogger } from '@eth-optimism/core-utils'
 import { Contract, Signer, ContractFactory } from 'ethers'
-import { DEFAULT_GAS_METER_PARAMS } from '@eth-optimism/rollup-core'
 
 /* Internal Imports */
 import {
@@ -13,6 +12,7 @@ import {
   makeAddressResolver,
   deployAndRegister,
   AddressResolverMapping,
+  DEFAULT_GAS_METER_PARAMS
 } from '../../test-helpers'
 
 /* Logging */
