+++ packages/contracts/test/contracts/ovm/execution-manager/ExecutionManager.code-opcodes.spec.ts
@@ -11,7 +11,6 @@ import {
   NULL_ADDRESS,
 } from '@eth-optimism/core-utils'
 import { Contract, ContractFactory, Signer } from 'ethers'
-import { DEFAULT_GAS_METER_PARAMS } from '@eth-optimism/rollup-core'
 
 /* Internal Imports */
 import {
@@ -24,6 +23,7 @@ import {
   makeAddressResolver,
   deployAndRegister,
   AddressResolverMapping,
+  DEFAULT_GAS_METER_PARAMS,
 } from '../../../test-helpers'
 
 /* Logging */
