+++ packages/contracts/test/contracts/ovm/execution-manager/ExecutionManager.create-opcodes.spec.ts
@@ -11,7 +11,6 @@ import {
 } from '@eth-optimism/core-utils'
 import { Contract, ContractFactory, Signer } from 'ethers'
 import { fromPairs } from 'lodash'
-import { DEFAULT_GAS_METER_PARAMS } from '@eth-optimism/rollup-core'
 
 /* Internal Imports */
 import {
@@ -23,6 +22,7 @@ import {
   makeAddressResolver,
   deployAndRegister,
   AddressResolverMapping,
+  DEFAULT_GAS_METER_PARAMS
 } from '../../../test-helpers'
 
 /* Logging */
