+++ packages/contracts/test/contracts/ovm/execution-manager/ExecutionManager.storage-opcodes.spec.ts
@@ -4,7 +4,6 @@ import '../../../setup'
 import { ethers } from '@nomiclabs/buidler'
 import { abi, getLogger, add0x, NULL_ADDRESS } from '@eth-optimism/core-utils'
 import { Contract, Signer, ContractFactory } from 'ethers'
-import { DEFAULT_GAS_METER_PARAMS } from '@eth-optimism/rollup-core'
 
 /* Internal Imports */
 import {
@@ -15,6 +14,7 @@ import {
   makeAddressResolver,
   deployAndRegister,
   AddressResolverMapping,
+  DEFAULT_GAS_METER_PARAMS,
 } from '../../../test-helpers'
 
 /* Logging */
