+++ packages/contracts/test/test-helpers/constants.ts
@@ -52,3 +52,15 @@ export const L2_TO_L1_MESSAGE_PASSER_OVM_ADDRESS =
 export const CHAIN_ID = 108
 export const ZERO_UINT = '00'.repeat(32)
 export const DEFAULT_FORCE_INCLUSION_PERIOD_SECONDS = 600
+
+const TX_FLAT_GAS_FEE = 30_000
+const MAX_SEQUENCED_GAS_PER_EPOCH = 2_000_000_000
+const MAX_QUEUED_GAS_PER_EPOCH = 2_000_000_000
+const GAS_RATE_LIMIT_EPOCH_LENGTH = 0
+export const DEFAULT_GAS_METER_PARAMS = [
+  TX_FLAT_GAS_FEE,
+  GAS_LIMIT,
+  GAS_RATE_LIMIT_EPOCH_LENGTH,
+  MAX_SEQUENCED_GAS_PER_EPOCH,
+  MAX_QUEUED_GAS_PER_EPOCH,
+]
