+++ packages/rollup-core/src/app/constants.ts
@@ -9,18 +9,6 @@ export const GAS_LIMIT = 1_000_000_000
 
 export const CHAIN_ID = 108
 
-const TX_FLAT_GAS_FEE = 30_000
-const MAX_SEQUENCED_GAS_PER_EPOCH = 2_000_000_000
-const MAX_QUEUED_GAS_PER_EPOCH = 2_000_000_000
-const GAS_RATE_LIMIT_EPOCH_LENGTH = 0
-export const DEFAULT_GAS_METER_PARAMS = [
-  TX_FLAT_GAS_FEE,
-  GAS_LIMIT,
-  GAS_RATE_LIMIT_EPOCH_LENGTH,
-  MAX_SEQUENCED_GAS_PER_EPOCH,
-  MAX_QUEUED_GAS_PER_EPOCH,
-]
-
 export const DEFAULT_UNSAFE_OPCODES: EVMOpcode[] = [
   Opcode.ADDRESS,
   Opcode.BALANCE,
