+++ packages/rollup-full-node/src/app/constants.ts
@@ -14,3 +14,15 @@ export const DEFAULT_OPCODE_WHITELIST_MASK =
 
 export const L2_TO_L1_MESSAGE_PASSER_OVM_ADDRESS =
   '0x4200000000000000000000000000000000000000'
+
+const TX_FLAT_GAS_FEE = 30_000
+const MAX_SEQUENCED_GAS_PER_EPOCH = 2_000_000_000
+const MAX_QUEUED_GAS_PER_EPOCH = 2_000_000_000
+const GAS_RATE_LIMIT_EPOCH_LENGTH = 0
+export const DEFAULT_GAS_METER_PARAMS = [
+  TX_FLAT_GAS_FEE,
+  GAS_LIMIT,
+  GAS_RATE_LIMIT_EPOCH_LENGTH,
+  MAX_SEQUENCED_GAS_PER_EPOCH,
+  MAX_QUEUED_GAS_PER_EPOCH,
+]
